# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sistema HRM - API
  description: Sistema de Gestión de Recursos Humanos y Nóminas
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: soporte@empresa.com
servers:
  - url: http://localhost:4000/api
    description: Servidor de desarrollo
  - url: https://tu-dominio.com/api
    description: Servidor de producción
...
paths:
...
  /employees/{id}/documents:
    post:
      tags: [Documentos]
      summary: Subir documento
      description: Sube un documento al expediente de un empleado
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID del empleado
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                document:
                  type: string
                  format: binary
                  description: Archivo a subir
                tags:
                  type: string
                  description: Etiquetas separadas por comas
                description:
                  type: string
                  description: Descripción del documento
      responses:
        '201':
          description: Documento subido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentWrapper'
        '413':
          description: Archivo demasiado grande
        '415':
          description: Tipo de archivo no permitido

    get:
      tags: [Documentos]
      summary: Listar documentos
      description: Retorna lista de documentos de un empleado
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Lista de documentos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListWrapper'
...
  /documents:
    get:
      tags: [Documentos]
      summary: Listar documentos
      description: Lista paginada y filtrable de documentos
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: 
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema: 
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: employeeId
          schema: 
            type: string
            format: uuid
          description: Filtrar por empleado
        - in: query
          name: q
          schema: 
            type: string
          description: Búsqueda por filename/description
        - in: query
          name: tags
          schema: 
            type: string
          description: Lista separada por comas
      responses:
        '200':
          description: Lista de documentos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListWrapper'

    post:
      tags: [Documentos]
      summary: Subir documento
      description: Sube un documento asociado a un empleado
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentUpload'
      responses:
        '201':
          description: Documento subido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentWrapper'
        '413':
          description: Archivo demasiado grande
        '415':
          description: Tipo de archivo no permitido
        '422':
          $ref: '#/components/responses/ValidationError'
  /documents/{id}:
    delete:
      tags: [Documentos]
      summary: Eliminar documento por ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: 
            type: string
            format: uuid
      responses:
        '200':
          description: Documento eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{id}/download:
    get:
      tags: [Documentos]
      summary: Descargar documento
      description: Devuelve el binario del archivo
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: 
            type: string
            format: uuid
      responses:
        '200':
          description: Archivo descargado
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Error de validación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
  schemas:
    MetaPagination:
      type: object
      properties:
        totalItems:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required: [totalItems, page, pageSize, totalPages]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
        meta:
          oneOf:
            - $ref: '#/components/schemas/MetaPagination'
            - type: object
      required: [success, data]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Error de validación"
            details:
              type: object
          required: [code, message]
      required: [success, error]
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            error:
              type: object
              properties:
                code:
                  type: string
                  example: "VALIDATION_ERROR"
                message:
                  type: string
                  example: "Error de validación"
                details:
                  type: object
                  description: Campo opcional con detalles específicos de los campos fallidos.
...
    Employee:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nationalId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        position:
          type: string
        hireDate:
          type: string
          format: date
        status:
          type: string
        department:
          $ref: '#/components/schemas/Department'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, nationalId, firstName, lastName, email, department, status, createdAt]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [ADMIN, RRHH, EMPLEADO]
        employee:
          $ref: '#/components/schemas/Employee'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, email, role, createdAt]
...
    Department:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
      required: [id, name]
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        tags:
          type: array
          items:
            type: string
        description:
          type: string
        uploadedBy:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
      required: [id, filename, mimeType, size, createdAt]
    DocumentWrapper:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Document'
    DocumentListWrapper:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Document'
            meta:
              $ref: '#/components/schemas/MetaPagination'
    DocumentUpload:
      type: object
      required: [document, employeeId]
      properties:
        document:
          type: string
          format: binary
        employeeId:
          type: string
          format: uuid
        description:
          type: string
        tags:
          type: string
          description: Etiquetas separadas por comas o JSON array como string
...